requirements.txt
pandas
numpy
transformers
accelerate
pyyaml
torch
yfinance
